// sitat-fe-workspace/projects/app-programmi/src/app/modules/models/interventi/interventi.model.ts
export interface InterventoBaseEntry {
    annoAvvio?: number;
    descrizione?: string;
    id?: number;
    idProgramma?: number;
    importoTotale?: number;
    inPianoAnnuale?: string;
    numProgetto?: number;
    numeroCui?: string;
}